Utilities

AnalogDataAssimilation.RMSE — Method

RMSE(e)

Returns the Root Mean Squared Error

AnalogDataAssimilation.ensure_pos_sym — Method

ensure_pos_sym(M; ϵ= 1e-8)

Ensure that matrix M is positive and symmetric to avoid numerical errors when numbers are small by doing (M + M')/2 + ϵ*I

reference : StateSpaceModels.jl

AnalogDataAssimilation.inv_using_SVD — Method

inv_using_SVD(Mat, eigvalMax)

SVD decomposition of Matrix.

AnalogDataAssimilation.mk_stochastic! — Method

mk_stochastic!(w)

Ensure the matrix is stochastic, i.e., the sum over the last dimension is 1.

AnalogDataAssimilation.normalise! — Method

normalise!( w )

Normalize the entries of a multidimensional array sum to 1.

AnalogDataAssimilation.resample! — Method

resample!( indx, w )

Multinomial resampler.

AnalogDataAssimilation.resample_multinomial — Method

resample_multinomial( w )

Multinomial resampler.

AnalogDataAssimilation.sample_discrete — Method

sample_discrete(prob, r, c)

Sampling from a non-uniform distribution.

AnalogDataAssimilation.sqrt_svd — Method

sqrt_svd(A)

Returns the square root matrix by SVD